refactoring the user module
