+++ src/main/java/br/com/caelum/carangobom/brand/Brand.java
@@ -19,7 +19,6 @@ public class Brand {
     private String name;
 
     public Brand() {
-
     }
 
     public Brand(String name) {
