+++ src/main/java/br/com/caelum/carangobom/user/User.java
@@ -1,6 +1,8 @@
 package br.com.caelum.carangobom.user;
 
 import br.com.caelum.carangobom.auth.Profile;
+import lombok.AllArgsConstructor;
+import lombok.Data;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
@@ -11,7 +13,10 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
+
 @Entity
+@Data
+@AllArgsConstructor
 public class User implements UserDetails {
 
     @Id
@@ -41,33 +46,6 @@ public class User implements UserDetails {
         this.password = password;
     }
 
-    public User() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return this.profile;
@@ -92,6 +70,4 @@ public class User implements UserDetails {
     public boolean isEnabled() {
         return true;
     }
-
-
 }
