+++ src/main/java/br/com/caelum/carangobom/user/UserDTO.java
@@ -1,30 +1,25 @@
 package br.com.caelum.carangobom.user;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Data
+@AllArgsConstructor
 public class UserDTO {
     private Long id;
     private String username;
 
     public UserDTO(User user) {
-        this.id = user.getId();
-        this.username = user.getUsername();
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public static List<UserDTO> convert(List<User> users) {
+    public static List<UserDTO> toUserList(List<User> users) {
         return users.stream().map(UserDTO::new).collect(Collectors.toList());
     }
 
-    public static UserDTO convertSingleUser(User user){
+    public static UserDTO toUser(User user){
         return new UserDTO(user);
     }
 }
