+++ src/main/java/br/com/caelum/carangobom/user/UserForm.java
@@ -1,9 +1,14 @@
 package br.com.caelum.carangobom.user;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
+@Data
+@AllArgsConstructor
 public class UserForm {
 
     @NotEmpty(message = "Usuário é um campo obrigatório")
@@ -16,31 +21,7 @@ public class UserForm {
     @Size(min = 3, max = 50, message = "A senha deve ser maior que {min} e menor que {max}")
     private String password;
 
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public UserForm() {
-    }
-
-    public UserForm(String username, String password) {
-        this.username = username;
-        this.password = password;
-    }
-
-    public User convert(){
+    public User toUser(){
         return new User(username, password);
     }
 }
